Split AnalyticsSuite into algorithms suites
